// src/UrlTracker.Backoffice.UI/api-client/types.gen.ts
export type RecommendationCollectionResponse = {
  results: Array<RecommendationResponse>;
  total: number;
};

export type RecommendationHistory = {
  firstOccurance: string;
  lastOccurance: string;
  averagePerDay: number;
  dailyOccurances: Array<DailyOccurance>;
  trend: RecomendationTrend;
};

export type RecommendationOrderBy = 'Importance' | 'MostRecentlyUpdated' | 'Url';

export type RecommendationResponse = {
  id: number;
  ignore: boolean;
  url: string;
  strategy: string;
  variableScore: number;
  updateDate: string;
  createDate: string;
};

export type RedactionScoreResponse = {
  key: string;
  score: number;
};

export type RedirectBulkRequest = {
  id: number;
  data: RedirectRequest;
};

export type RedirectCollectionResponse = {
  results: Array<RedirectResponse>;
  total: number;
};

export type RedirectRequest = {
  source: StrategyViewModel;
  target: StrategyViewModel;
  permanent: boolean;
  retainQuery: boolean;
  force: boolean;
  advanced: boolean;
};
// ...
export type StrategyViewModel = {
  strategy: string;
  value: string;
};